backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["yakimanjusuki <yakimanjusuki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^20.1.0"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
langchain = "^0.0.132"
python-dotenv = "^1.0.0"
openai = "^0.27.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
